+++ mods/tiamot_default_world/biomes/kelp_forest.lua
@@ -157,7 +157,7 @@ tdw.build_biome(ID, function(ctx)
     local entries = {
         { code = 1, to = 4 * km, material = blocks.sand },
         { code = 2, to = 1 * km, material = blocks.ocean_moss },
-        { code = 2, from = 1 * km, to = 40 * km, material = blocks.stone },
+        { code = 2, from = 1 * km, to = 12 * km, material = blocks.stone },
         { code = 3, to = 3 * km, material = blocks.gravel },
     }
     local fills = {